+++ app/assets/javascripts/widgets/lightbox.js
@@ -28,9 +28,12 @@ jQuery.fn.center = (function() {
     this.subscribe("widget/ready", function(evt) {
       $.extend(self, {
         lightbox: $("#lightbox"),
+        navigation: $("#lightbox-navigation"),
         imageset: $("#lightbox-imageset"),
         backdrop: $("#lightbox-backdrop"),
         closelink: $("#lightbox-close-link"),
+        scrollleft: $("#lightbox-scrollleft"),
+        scrollright: $("#lightbox-scrollright"),
         image: $("#lightbox-image"),
         body: $(document.body),
         window: $(window)
@@ -47,12 +50,25 @@ jQuery.fn.center = (function() {
         evt.preventDefault();
         self.resetLightbox();
       });
-      self.backdrop.click(self.resetLightbox);
       self.lightbox.click(self.resetLightbox);
 
       self.backdrop.click(function(evt) {
         evt.preventDefault();
         self.resetLightbox();
+      });
+
+      self.scrollleft.click(function(evt){
+        evt.preventDefault();
+        evt.stopPropagation();
+        self.navigation.animate({scrollLeft: (self.navigation.scrollLeft()
+           - (self.window.width() - 150))}, 200, 'swing');
+      });
+
+      self.scrollright.click(function(evt){
+        evt.preventDefault();
+        evt.stopPropagation();
+        self.navigation.animate({scrollLeft: (self.navigation.scrollLeft()
+           + (self.window.width() - 150))}, 200, 'swing');
       });
 
       self.body.keydown(function(evt) {
@@ -122,6 +138,8 @@ jQuery.fn.center = (function() {
       self
         .selectImage(imageThumb)
         .revealLightbox();
+
+      self.scrollToThumbnail(imageThumb);
     };
 
     this.imagesetImageClicked = function(evt) {
@@ -131,10 +149,17 @@ jQuery.fn.center = (function() {
       self.selectImage($(this));
     };
 
+    this.scrollToThumbnail = function(imageThumb) {
+      self.navigation.animate({scrollLeft: (self.navigation.scrollLeft()
+         + imageThumb.offset().left +35 - (self.window.width() / 2))}, 200, 'swing');
+    }
+
     this.selectImage = function(imageThumb) {
       $(".selected", self.imageset).removeClass("selected");
       imageThumb.addClass("selected");
       self.image.attr("src", imageThumb.attr("data-full-photo"));
+
+      self.scrollToThumbnail(imageThumb);
 
       return self;
     };